add permissions to delete shoutout

+++ backend/src/API/shoutoutAPI.ts
@@ -48,5 +48,11 @@ export const deleteShoutout = async (uuid: string, user: IdolMember): Promise<vo
   if (!shoutout) {
     throw new NotFoundError(`No shoutout with id '${uuid}' found.`);
   }
+  const isLeadOrAdmin = await PermissionsManager.isLeadOrAdmin(user);
+  if (!isLeadOrAdmin || shoutout.giver.email !== user.email) {
+    throw new PermissionError(
+      `User with email: ${user.email} can't delete a shoutout from a different user!`
+    );
+  }
   await ShoutoutsDao.deleteShoutout(uuid);
 };
